Not beautiful, but it gets the job done. Eventually more failsave checks need to be implemented.

+++ MakeMacImage.sh
@@ -2,43 +2,88 @@
 
 TMPDIR=/tmp
 RELEASE="11"
+
 SETFILE=/Developer/Tools/SetFile
 HDIUTIL=/usr/bin/hdiutil
+SED=/usr/bin/sed
+CUT=/usr/bin/cut
+GREP=/usr/bin/grep
+STRIP=/usr/bin/strip
+
+EXAMPLES="brusselator.cps"
 
 # check if the binary is there
 if test -e ./copasi/CopasiUI/CopasiUI.app/Contents/MacOS/CopasiUI ; then
-    echo -n "Please build the copasi executable first.\n"
-    exit 1;
-else
-
+# test if the binary is statically linked agains libqt
+    if otool -L ./copasi/CopasiUI/CopasiUI.app/Contents/MacOS/CopasiUI | grep libqt 2>&1 > /dev/null ;then
+        echo ""
+        echo "CopasiUI is dynamicall linked against Qt."
+        echo "Please link the binary statically before making an image."
+        echo ""
+        exit 1
+    fi
+# strip the binary
+    echo "Removing debugging symbols from the binary."
+    ${STRIP} ./copasi/CopasiUI/CopasiUI.app/Contents/MacOS/CopasiUI  
+    
 # make a temporary directory and copy the CopasiUI.app directory there.
+    echo "Creating temporary directory."
     mkdir ${TMPDIR}/copasi;
+    echo "Copying application bundle."
     cp -r copasi/CopasiUI/CopasiUI.app ${TMPDIR}/copasi/
 
 # copy the icon into the Resources directory
-    cp copasi.icns ${TMPDIR}/copasi/CopasiUI/CopasiUI.app/Contents/Resources/
+    echo "Creating Resources directory."
+    mkdir -p ${TMPDIR}/copasi/CopasiUI.app/Contents/Resources
+    echo "Copy the icon file to the bundle."
+    cp copasi.icns ${TMPDIR}/copasi/CopasiUI.app/Contents/Resources/
 
 # set the icon in the Info.plist file in the Contents directory
-
+# the icon file name is located between <string></string> tags in the line after
+# <key>CFBundleIconFile</key>
+# Maybe this would better be done with a dom parser.
+echo "Set the icon in the Info.plist file."
+    LINE=`${GREP} -n "CFBundleIconFile" ${TMPDIR}/copasi/CopasiUI.app/Contents/Info.plist | cut -c1`
+    LINE=$((LINE+1))
+    ${SED} -e "${LINE}s%<string>.*</string>%<string>copasi.icns</string>%" ${TMPDIR}/copasi/CopasiUI.app/Contents/Info.plist > ${TMPDIR}/tmp.plist
+    mv ${TMPDIR}/tmp.plist ${TMPDIR}/copasi/CopasiUI.app/Contents/Info.plist
 
 # copy the examples into the Resources directory
-    mkdir ${TMPDIR}/copasi/CopasiUI.app/Contents/Resources/EXAMPLES
-
+    echo "Make example directory."
+    mkdir -p ${TMPDIR}/copasi/CopasiUI.app/Contents/Resources/EXAMPLES
+    echo "Copy examples to example directory."
+    for EXAMPLE in $EXAMPLES;do
+        cp ./gps/${EXAMPLE} ${TMPDIR}/copasi/CopasiUI.app/Contents/Resources/EXAMPLES/
+    done
 
 # copy the files for the wizard into the Resources directory
+    echo "Making directory for wizard."
     mkdir -p ${TMPDIR}/copasi/CopasiUI.app/Contents/Resources/doc/html
-    cp ./copasi/wizard/help_html/*.html ${TMPDIR}/copasi/CopasiUI.app/Contents/Resources/doc/
-    cp -r ./copasi/wizard/help_html/figures ${TMPDIR}/copasi/CopasiUI.app/Contents/Resources/doc/
+    echo "Copying files for wizard."
+    cp ./copasi/wizard/help_html/*.html ${TMPDIR}/copasi/CopasiUI.app/Contents/Resources/doc/html/
+    cp -r ./copasi/wizard/help_html/figures ${TMPDIR}/copasi/CopasiUI.app/Contents/Resources/doc/html/
 
 # add the readme to the image
+    echo "Copying readme file."
     cp README_MAC.rtf ${TMPDIR}/copasi/
 
 # set the bundles flag with SetFile
+    echo "Setting the bundle flag."
     ${SETFILE} -a b ${TMPDIR}/copasi/CopasiUI.app
 
 # make an image with hdiutil from the folder
-    ${HDIUTIL} create -size ${SIZE}k -format UDZ0 -srcfolder ${TMPDIR}/copasi -volname Copasi-${RELEASE} -uid 99 -gid 99 -scrub ${TMPDIR}/Copasi-${RELEASE}-Mac.dmg 
+    echo "Creating compressed image."
+    ${HDIUTIL} create -srcfolder ${TMPDIR}/copasi  -uid 99 -gid 99 -scrub -volname Copasi-${RELEASE} ${TMPDIR}/Copasi-${RELEASE}-Mac.dmg 
 
-
+# move the image to this directory and delete the temporary diretory
+    echo "Cleaning up in temporary directory."
+    mv ${TMPDIR}/Copasi-${RELEASE}-Mac.dmg .
+    rm -rf ${TMPDIR}/copasi
+ else
+    echo ""
+    echo "Please build the copasi executable first."
+    echo ""
+    exit 1;
+   
 fi
